+++ src/stateMachine/mutation/BQL/parse.ts
@@ -159,7 +159,7 @@ export const parseBQLMutation = async (
 					...(value.$thingType && { $thingType: value.$thingType }),
 					...shake(pick(value, dataFieldPaths || [''])),
 					$op: getChildOp() as BormOperation,
-					$bzId: value.$tempId ? value.$tempId : value.$bzId,
+					$bzId: value.$bzId,
 				};
 
 				/// split nodes with multiple ids // why? //no longer doing that
@@ -196,10 +196,6 @@ export const parseBQLMutation = async (
 						throw new Error('No parent id found');
 					}
 
-					if (value[Symbol.for('relation') as any] === '$self') {
-						return;
-					}
-
 					const getLinkObjOp = () => {
 						if (value.$op === 'delete') {
 							if (ownRelation) {
@@ -226,6 +222,17 @@ export const parseBQLMutation = async (
 						}
 						return 'match';
 					};
+					//validate that field is an actual role from the relation
+					const relationSchema = getCurrentSchema(schema, {
+						$thing: edgeSchema.relation,
+						$thingType: 'relation',
+					}) as EnrichedBormRelation;
+					const roles = Object.keys(relationSchema.roles);
+					if (!roles.includes(edgeSchema.plays)) {
+						throw new Error(
+							`[Wrong format] Field ${edgeSchema.plays} is not a role of relation ${edgeSchema.relation}`,
+						);
+					}
 
 					const edgeType1 = {
 						$bzId: linkTempId,