+++ src/stateMachine/mutation/machine.ts
@@ -1,4 +1,3 @@
-import { createMachine, interpret, invoke, transition, reduce, state as final, guard } from 'robot3';
 import type {
 	BQLMutation,
 	BQLMutationBlock,
@@ -16,6 +15,9 @@ import { parseBQLMutation } from './BQL/parse';
 import { buildTQLMutation } from './TQL/build';
 import { mutationPreQuery } from './BQL/preQuery';
 
+import { createMachine, invoke, transition, reduce, guard, interpret, state } from './robot3-wrapper';
+
+const final = state;
 type MachineContext = {
 	bql: {
 		raw: BQLMutationBlock | BQLMutationBlock[];
@@ -92,8 +94,8 @@ const updateTQLRes = (ctx: MachineContext, event: any) => {
 
 const enrich = async (ctx: MachineContext) => {
 	return Object.keys(ctx.bql.current).length
-		? enrichBQLMutation(ctx.bql.current, ctx.schema)
-		: enrichBQLMutation(ctx.bql.raw, ctx.schema);
+		? enrichBQLMutation(ctx.bql.current, ctx.schema, ctx.config)
+		: enrichBQLMutation(ctx.bql.raw, ctx.schema, ctx.config);
 };
 
 const preQuery = async (ctx: MachineContext) => {
@@ -161,13 +163,16 @@ export const machine = createMachine(
 		success: final(),
 		error: final(),
 	},
+	//@ts-expect-error - todo
 	(ctx: MachineContext) => ctx,
 );
 
 export const awaitMachine = async (context: MachineContext) => {
 	return new Promise<MachineContext>((resolve, reject) => {
+		//@ts-expect-error - todo
 		interpret(
 			machine,
+			//@ts-expect-error - todo
 			(service) => {
 				if (service.machine.state.name === 'success') {
 					resolve(service.context);