+++ src/types/config/base.ts
@@ -12,6 +12,7 @@ export type MutationConfig = {
 	noMetadata?: boolean;
 	preQuery?: boolean;
 	ignoreNonexistingThings?: boolean;
+	context?: Record<string, any>;
 };
 
 export type BormConfig = {